[Refactor][Kernels] give every family one call record base
``Op.select_kernel_key`` asks each implementation ``supports(call)``, and
``Kernel.supports`` reads one field: ``arch``. Resolving it — and rendering the
record when selection fails — belonged in one place before a second family
needed a record of its own, or each would carry its own copy of the same probe.

``CallSpec`` holds the device fields and their lazy read; ``AttentionCall``
subclasses it and keeps only its semantic fields. No behaviour changes: the
device is still read when the call is made rather than at construction, a stated
``arch`` still wins, and the golden table is untouched.
